from home, change flaot in decimal in database and bug fix in position calculation

+++ Antea25/Models/GpsPosition.cs
@@ -10,8 +10,8 @@ namespace Antea25.Models
         public int GpsPositionId { get; set; }
         public string UserId { get; set; }
         public int DeviceId { get; set; }
-        public float GpsPositionLatitude { get; set; }
-        public float GpsPositionLongitude { get; set; }
+        public decimal GpsPositionLatitude { get; set; }
+        public decimal GpsPositionLongitude { get; set; }
         public DateTime GpsPositionDate { get; set; }
 
         public string GpsPositionLatitudeRaw { get; set; }
